move tools into sub folder
