+++ code/python_tooling/dashboard/server.py
@@ -10,6 +10,5 @@ app = Dash(title='Reprojection', update_title=None)
 #  trivial but can be done with some tk tools or other libraries
 DB_DIR = '../../test_data/'
 
-
 # TODO(Jack): Place meta data like this in config file or in database. For now we use globals...
 IMAGE_DIMENSIONS = (512, 512)
