+++ code/python_tooling/dashboard/tools/plot_pose_figure.py
@@ -1,6 +1,6 @@
 import plotly.graph_objects as go
 
-from dashboard.time_handling import timestamps_to_elapsed_seconds, calculate_ticks_from_timestamps
+from dashboard.tools.time_handling import timestamps_to_elapsed_seconds, calculate_ticks_from_timestamps
 
 
 # NOTE(Jack): Think about it this way. The moment that we have two separate arrays we cannot/should not ever sort them.
