+++ code/python_tooling/setup.py
@@ -5,8 +5,7 @@ setup(
     version="0.0.0",
     author="Jack Borer",
     author_email="reprojection.calibration@gmail.com",
-    description=(
-        "A package for handling and visualization of calibration data."),
+    description="A package for handling and visualization of calibration data.",
     license="GNU AGPL",
     url="https://github.com/reprojection-calibration/reprojection",
     packages=['dashboard', 'database', 'generated', 'tests'],
