+++ code/python_tooling/tests/test_dashboard_time_handling.py
@@ -1,6 +1,6 @@
 import unittest
 
-from dashboard.time_handling import calculate_ticks_from_timestamps, extract_timestamps_and_poses_sorted, \
+from dashboard.tools.time_handling import calculate_ticks_from_timestamps, extract_timestamps_and_poses_sorted, \
     timestamps_to_elapsed_seconds
 from database.types import PoseType
